:fire: remove unused bootstrap file
This `dotfiles` bootstrap file will be moved to a gist instead and
linked here once ready. The point of this file is to bootstrap our
dotfiles, so having the function set as a fish function makes little
practical sense. A `fish` script will instead be provided that downloads
the repo and sets it up correctly.